backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.4
